[514] BSR : Fix les entreprises s'engagent qui ne respectent pas les standards
