+++ src/client/components/features/LesEntreprisesSEngagent/Rejoignez/Inscription/RejoignezMobilisationInscription.module.scss
@@ -1,4 +1,4 @@
-@import "@styles/utilities";
+@import "@styles/_utilities";
 
 .header {
   background-color: $color-on-primary;
