+++ src/client/components/features/LesEntreprisesSEngagent/Rejoignez/RejoignezMobilisation.module.scss
@@ -1,4 +1,4 @@
-@import "@styles/utilities";
+@import "@styles/_utilities";
 
 .content {
   padding: 32px;
