+++ src/client/services/les-entreprises-s-engagent/lesEntreprisesSEngagent.service.ts
@@ -1,10 +1,11 @@
+import { FormulaireEngagement } from '~/client/components/features/LesEntreprisesSEngagent/Rejoignez/Inscription/Inscription';
 import { HttpClientService } from '~/client/services/httpClient.service';
-import { FormulaireEngagement } from '~/pages/les-entreprises-s-engagent/inscription';
 import { createFailure, createSuccess, Either } from '~/server/errors/either';
 import { ErreurMétier } from '~/server/errors/erreurMétier.types';
 
 export class LesEntreprisesSEngagentService {
-  constructor(private httpClientService: HttpClientService) {}
+  constructor(private httpClientService: HttpClientService) {
+  }
 
   async envoyerFormulaireEngagement(formulaire: FormulaireEngagement): Promise<Either<void>> {
     try {
